+++ src/nile-rs/core/mod.rs
@@ -1,3 +1,5 @@
 pub mod accounts;
-mod deployments;
+pub mod deployments;
 pub use deployments::Deployments;
+pub mod status;
+pub mod types;