+++ src/nile-rs/lib.rs
@@ -1,4 +1,5 @@
 pub mod common;
 pub mod config;
 pub mod core;
+pub mod nre;
 pub mod utils;